Route Format  is now Semantic way
