+++ components/course/course-section.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Course } from "@/types/course.schema";
 import React, { FunctionComponent } from "react";
-import CourseCard from "../course-card";
+import CourseCard from "../common/course-card";
 import SearchCourse from "./search-course";
 import { useSearchParams } from "next/navigation";
 
@@ -15,14 +15,14 @@ const CourseSection: FunctionComponent<ICourseSection> = ({ courses }) => {
     <div className="space-y-10">
       <SearchCourse />
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
-        {courses
-          .filter((course) => {
-            if (!search) return true;
-            return course.name.toLowerCase().includes(search.toLowerCase());
-          })
-          .map((course) => (
-            <CourseCard key={course.id} course={course} />
-          ))}
+        {courses?.length
+          ? courses
+              .filter((course) => {
+                if (!search) return true;
+                return course.name.toLowerCase().includes(search.toLowerCase());
+              })
+              .map((course) => <CourseCard key={course.id} course={course} />)
+          : null}
       </div>
     </div>
   );
