+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^4.11.0",
+    "sharp": "^0.32.6",
     "swiper": "^11.0.3",
     "zod": "^3.22.4"
   },
